% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mhq}
\alias{mhq}
\title{Michigan Hand Outcomes Questionnaire (MHQ) Data}
\format{
A 5 x 63 \link[dplyr:reexports]{dplyr::tibble}
\itemize{
\item hand. The injured hand
\itemize{
\item 1, right | 2, left | 3, both
}
\item I. The following questions refer to the function of your hand(s)/wrist(s) during the past week.
\itemize{
\item A. The following questions refer to your right hand/wrist.
\itemize{
\item ia1. Overall, how well did your right hand work?
\itemize{
\item 1, Very good | 2, Good | 3, Fair | 4, Poor | 5, Very poor
}
\item ia2. How well did your right fingers move?
\itemize{
\item 1, Very good | 2, Good | 3, Fair | 4, Poor | 5, Very poor
}
\item ia3. How well did your right wrist move?
\itemize{
\item 1, Very good | 2, Good | 3, Fair | 4, Poor | 5, Very poor
}
\item ia4. How was the strength in your right hand?
\itemize{
\item 1, Very good | 2, Good | 3, Fair | 4, Poor | 5, Very poor
}
\item ia5. How was the sensation (feeling) in your right hand?
\itemize{
\item 1, Very good | 2, Good | 3, Fair | 4, Poor | 5, Very poor
}
}
\item B The following questions refer to your left hand/wrist.
\itemize{
\item ib1. Overall, how well did your left hand work?
\itemize{
\item 1, Very Good | 2, Good | 3, Fair | 4, Poor | 5, Very Poor
}
\item ib2. How well did your left fingers move?
\itemize{
\item 1, Very Good | 2, Good | 3, Fair | 4, Poor | 5, Very Poor
}
\item ib3. How well did your left wrist move?
\itemize{
\item 1, Very Good | 2, Good | 3, Fair | 4, Poor | 5, Very Poor
}
\item ib4. How was the strength in your left hand?
\itemize{
\item 1, Very Good | 2, Good | 3, Fair | 4, Poor | 5, Very Poor
}
\item ib5. How was the sensation (feeling) in your left hand?
\itemize{
\item 1, Very Good | 2, Good | 3, Fair | 4, Poor | 5, Very Poor
}
}
}
\item II. The following questions refer to the ability of your hand(s) to do certain tasks during the past week.
\itemize{
\item A. How difficult was it for you to perform the following activities using your right hand?
\itemize{
\item iia1. Turn a door knob
\itemize{
\item 1, Not at all difficult | 2, A little difficult | 3, Somewhat difficult | 4, Moderately difficult | 5, Very difficult
}
\item iia2. Pick up a coin
\itemize{
\item 1, Not at all difficult | 2, A little difficult | 3, Somewhat difficult | 4, Moderately difficult | 5, Very difficult
}
\item iia3. Hold a glass of water
\itemize{
\item 1, Not at all difficult | 2, A little difficult | 3, Somewhat difficult | 4, Moderately difficult | 5, Very difficult
}
\item iia4. Turn a key in a lock
\itemize{
\item 1, Not at all difficult | 2, A little difficult | 3, Somewhat difficult | 4, Moderately difficult | 5, Very difficult
}
\item iia5. Hold a frying pan
\itemize{
\item 1, Not at all difficult | 2, A little difficult | 3, Somewhat difficult | 4, Moderately difficult | 5, Very difficult
}
}
\item B. How difficult was it for you to perform the following activities using your left hand?
\itemize{
\item iib1. Turn a door knob
\itemize{
\item 1, Not at all difficult | 2, A little difficult | 3, Somewhat difficult | 4, Moderately difficult | 5, Very difficult
}
\item iib2. Pick up a coin
\itemize{
\item 1, Not at all difficult | 2, A little difficult | 3, Somewhat difficult | 4, Moderately difficult | 5, Very difficult
}
\item iib3. Hold a glass of water
\itemize{
\item 1, Not at all difficult | 2, A little difficult | 3, Somewhat difficult | 4, Moderately difficult | 5, Very difficult
}
\item iib4. Turn a key in a lock
\itemize{
\item 1, Not at all difficult | 2, A little difficult | 3, Somewhat difficult | 4, Moderately difficult | 5, Very difficult
}
\item iib5. Hold a frying pan
\itemize{
\item 1, Not at all difficult | 2, A little difficult | 3, Somewhat difficult | 4, Moderately difficult | 5, Very difficult
}
}
\item C. How difficult was it for you to perform the following activities using both of your hands?
\itemize{
\item iic1. Open a jar
\itemize{
\item 1, Not at all difficult | 2, A little difficult | 3, Somewhat difficult | 4, Moderately difficult | 5, Very difficult
}
\item iic2. Button a shirt/blouse
\itemize{
\item 1, Not at all difficult | 2, A little difficult | 3, Somewhat difficult | 4, Moderately difficult | 5, Very difficult
}
\item iic3. Eat with a knife/fork
\itemize{
\item 1, Not at all difficult | 2, A little difficult | 3, Somewhat difficult | 4, Moderately difficult | 5, Very difficult
}
\item iic4. Carry a grocery bag
\itemize{
\item 1, Not at all difficult | 2, A little difficult | 3, Somewhat difficult | 4, Moderately difficult | 5, Very difficult
}
\item iic5. Wash dishes
\itemize{
\item 1, Not at all difficult | 2, A little difficult | 3, Somewhat difficult | 4, Moderately difficult | 5, Very difficult
}
\item iic6. Wash your hair
\itemize{
\item 1, Not at all difficult | 2, A little difficult | 3, Somewhat difficult | 4, Moderately difficult | 5, Very difficult
}
\item iic7. Tie shoelaces/knots
\itemize{
\item 1, Not at all difficult | 2, A little difficult | 3, Somewhat difficult | 4, Moderately difficult | 5, Very difficult
}
}
}
\item III. The following questions refer to how you did in your normal work (including both housework and school work) during the past 4 weeks.
\itemize{
\item iii1. How often were you unable to do your work because of problems with your hand(s)/wrist(s)?
\itemize{
\item 1, Always | 2, Often | 3, Sometimes | 4, Rarely | 5, Never
}
\item iii2. How often did you have to shorten your work day because of problems with your hand(s)/wrist(s)?
\itemize{
\item 1, Always | 2, Often | 3, Sometimes | 4, Rarely | 5, Never
}
\item iii3. How often did you have to take it easy at your work because of problems with yoru hand(s)/wrist(s)?
\itemize{
\item 1, Always | 2, Often | 3, Sometimes | 4, Rarely | 5, Never
}
\item iii4. How often did you accomplish less in your work because of problems with your hand(s)/wrist(s)?
\itemize{
\item 1, Always | 2, Often | 3, Sometimes | 4, Rarely | 5, Never
}
\item iii5. How often did you take longer to do the tasks in your work because of problems with your hand(s)/wrist(s)?
\itemize{
\item 1, Always | 2, Often | 3, Sometimes | 4, Rarely | 5, Never
}
}
\item IV. The following questions refer to how much pain you had in your hand(s)/wrist(s) during the past week.
\itemize{
\item A. The following questions refer to your right hand.
\itemize{
\item iva1. How often did you have pain in your right hand/wrist?
\itemize{
\item 1, Always | 2, Often | 3, Sometimes | 4, Rarely | 5, Never
}
\item iva2. Please describe the pain you have in your right hand/wrist.
\itemize{
\item 1, Always | 2, Often | 3, Sometimes | 4, Rarely | 5, Never
}
\item iva3. How often did the pain in your right hand/wrist interfere with your sleep?
\itemize{
\item 1, Always | 2, Often | 3, Sometimes | 4, Rarely | 5, Never
}
\item iva4. How often did the pain in your right hand/wrist interfere with your daily activities (such as eating or bathing)?
\itemize{
\item 1, Always | 2, Often | 3, Sometimes | 4, Rarely | 5, Never
}
\item iva5. How often did the pain in your right hand/wrist make you unhappy?
\itemize{
\item 1, Always | 2, Often | 3, Sometimes | 4, Rarely | 5, Never
}
}
\item B. The following questions refer to your left hand.
\itemize{
\item iva1. How often did you have pain in your left hand/wrist?
\itemize{
\item 1, Always | 2, Often | 3, Sometimes | 4, Rarely | 5, Never
}
\item iva2. Please describe the pain you have in your left hand/wrist.
\itemize{
\item 1, Always | 2, Often | 3, Sometimes | 4, Rarely | 5, Never
}
\item iva3. How often did the pain in your left hand/wrist interfere with your sleep?
\itemize{
\item 1, Always | 2, Often | 3, Sometimes | 4, Rarely | 5, Never
}
\item iva4. How often did the pain in your left hand/wrist interfere with your daily activities (such as eating or bathing)?
\itemize{
\item 1, Always | 2, Often | 3, Sometimes | 4, Rarely | 5, Never
}
\item iva5. How often did the pain in your left hand/wrist make you unhappy?
\itemize{
\item 1, Always | 2, Often | 3, Sometimes | 4, Rarely | 5, Never
}
}
}
\item V. The following questions refer to the appearance (look) of your hands during the past week.
\itemize{
\item A. The following questions refer to your right hand.
\itemize{
\item va1. I was satisfied with the appearance (look) of my right hand.
\itemize{
\item 1, Strongly agree, | 2, Agree | 3, Neither agree nor disagree | 4, Disagree | 5, Strongly disagree
}
\item va2. The appearance (look) of my right hand sometimes made me uncomfortable in public.
\itemize{
\item 1, Strongly agree, | 2, Agree | 3, Neither agree nor disagree | 4, Disagree | 5, Strongly disagree
}
\item va3. The appearance (look) of my right hand made me depressed.
\itemize{
\item 1, Strongly agree, | 2, Agree | 3, Neither agree nor disagree | 4, Disagree | 5, Strongly disagree
}
\item va4. The appearance (look) of my right hand interfered with my normal social activities.
\itemize{
\item 1, Strongly agree, | 2, Agree | 3, Neither agree nor disagree | 4, Disagree | 5, Strongly disagree
}
}
\item B. The following questions refer to your left hand.
\itemize{
\item vb1. I was satisfied with the appearance (look) of my left hand.
\itemize{
\item 1, Strongly agree, | 2, Agree | 3, Neither agree nor disagree | 4, Disagree | 5, Strongly disagree
}
\item vb2. The appearance (look) of my left hand sometimes made me uncomfortable in public.
\itemize{
\item 1, Strongly agree, | 2, Agree | 3, Neither agree nor disagree | 4, Disagree | 5, Strongly disagree
}
\item vb3. The appearance (look) of my left hand made me depressed.
\itemize{
\item 1, Strongly agree, | 2, Agree | 3, Neither agree nor disagree | 4, Disagree | 5, Strongly disagree
}
\item vb4. The appearance (look) of my left hand interfered with my normal social activities.
\itemize{
\item 1, Strongly agree, | 2, Agree | 3, Neither agree nor disagree | 4, Disagree | 5, Strongly disagree
}
}
}
\item VI. The following questions refer to your satisfaction with your hand(s)/wrist(s) during the past week.
\itemize{
\item A. The following questions refer to your right hand/wrist.
\itemize{
\item via1. Overall function of your right hand
\itemize{
\item 1, Very satisfied | 2, Somewhat satisfied | 3, Neither satisfied nor dissatisfied | 4, Somewhat dissatisfied | 5, Very dissatisfied
}
\item via2. Motion of the fingers in your right hand
\itemize{
\item 1, Very satisfied | 2, Somewhat satisfied | 3, Neither satisfied nor dissatisfied | 4, Somewhat dissatisfied | 5, Very dissatisfied
}
\item via3. Motion of your right wrist
\itemize{
\item 1, Very satisfied | 2, Somewhat satisfied | 3, Neither satisfied nor dissatisfied | 4, Somewhat dissatisfied | 5, Very dissatisfied
}
\item via4. Strength of your right hand
\itemize{
\item 1, Very satisfied | 2, Somewhat satisfied | 3, Neither satisfied nor dissatisfied | 4, Somewhat dissatisfied | 5, Very dissatisfied
}
\item via5. Pain level of your right hand
\itemize{
\item 1, Very satisfied | 2, Somewhat satisfied | 3, Neither satisfied nor dissatisfied | 4, Somewhat dissatisfied | 5, Very dissatisfied
}
\item via6. Sensation (feeling) of your right hand
\itemize{
\item 1, Very satisfied | 2, Somewhat satisfied | 3, Neither satisfied nor dissatisfied | 4, Somewhat dissatisfied | 5, Very dissatisfied
}
}
\item B The following questions refer to your left hand/wrist.
\itemize{
\item vib1. Overall function of your left hand
\itemize{
\item 1, Very satisfied | 2, Somewhat satisfied | 3, Neither satisfied nor dissatisfied | 4, Somewhat dissatisfied | 5, Very dissatisfied
}
\item vib2. Motion of the fingers in your left hand
\itemize{
\item 1, Very satisfied | 2, Somewhat satisfied | 3, Neither satisfied nor dissatisfied | 4, Somewhat dissatisfied | 5, Very dissatisfied
}
\item vib3. Motion of your left wrist
\itemize{
\item 1, Very satisfied | 2, Somewhat satisfied | 3, Neither satisfied nor dissatisfied | 4, Somewhat dissatisfied | 5, Very dissatisfied
}
\item vib4. Strength of your left hand
\itemize{
\item 1, Very satisfied | 2, Somewhat satisfied | 3, Neither satisfied nor dissatisfied | 4, Somewhat dissatisfied | 5, Very dissatisfied
}
\item vib5. Pain level of your left hand
\itemize{
\item 1, Very satisfied | 2, Somewhat satisfied | 3, Neither satisfied nor dissatisfied | 4, Somewhat dissatisfied | 5, Very dissatisfied
}
\item vib6. Sensation (feeling) of your left hand
\itemize{
\item 1, Very satisfied | 2, Somewhat satisfied | 3, Neither satisfied nor dissatisfied | 4, Somewhat dissatisfied | 5, Very dissatisfied
}
}
}
}
}
\usage{
mhq
}
\description{
Example Michigan Hand Outcomes Questionnaire (MHQ) Data
}
\keyword{datasets}
