% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aodl.R
\name{mhq_aodl}
\alias{mhq_aodl}
\title{MHQ Activities of Daily Living}
\usage{
mhq_aodl(
  data,
  ...,
  handinj,
  score = c("raw", "mean", "stnd", "norm"),
  norm_m = 0,
  norm_s = 1
)
}
\arguments{
\item{data}{a \link[dplyr:reexports]{dplyr::tibble} containing the (sub)scale item responses to be
scored}

\item{...}{<\link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select}> columns of the (sub)scale items in
order (See Note).}

\item{handinj}{<\link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select}> column specification of the
affected hand(s)}

\item{score}{desired score class}

\item{norm_m}{mean value to use when computing normative scores}

\item{norm_s}{standard deviation value to use when computing normative scores}
}
\value{
data augmented with score columns:
\enumerate{
\item mhq_aodl_left_\if{html}{\out{<score>}}
\itemize{
\item score from the left hand items
}
\item mhq_aodl_right_\if{html}{\out{<score>}}
\itemize{
\item score from the right hand items
}
\item mhq_aodl_both_\if{html}{\out{<score>}}
\itemize{
\item score from the two hand items
}
\item mhq_aodl_overall_left_\if{html}{\out{<score>}}
\itemize{
\item an average of the left and two hand scores
}
\item mhq_aodl_overall_right_\if{html}{\out{<score>}}
\itemize{
\item an average of the right and two hand scores
}
\item mhq_aodl_overall_both_\if{html}{\out{<score>}}
\itemize{
\item an average of the left, right, and two hand scores
}
\item mhq_aodl_\if{html}{\out{<score>}}
\itemize{
\item the requested activities of daily living score
}
}
}
\description{
MHQ Activities of Daily Living
}
\note{
Items indicated in the ellipsis (...) are expected to be in the following
order:

II The following questions refer to the ability of your hand(s) to do
certain tasks during the past week.

A How difficult was it for you to perform the following activities using
your right hand?
\enumerate{
\item Turn a door knob
\item Pick up a coin
\item Hold a glass of water
\item Turn a key in a lock
\item Hold a frying pan
}

B How difficult was it for you to perform the following activities using
your left hand?
6. Turn a door knob
7. Pick up a coin
8. Hold a glass of water
9. Turn a key in a lock
10. Hold a frying pan

C How difficult was it for you to perform the following activities using
both of your hands?
11. Open a jar
12. Button a shirt/blouse
13. Eat with a knife/fork
14. Carry a grocery bag
15. Wash dishes
16. Wash your hair
17. Tie shoelaces/knots
}
\examples{
mhq_aodl(mhq, iia1:iic7, handinj = handinj, score = "stnd")

}
