% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asth.R
\name{mhq_asth}
\alias{mhq_asth}
\title{MHQ Aesthetics}
\usage{
mhq_asth(
  data,
  ...,
  handinj,
  score = c("raw", "mean", "stnd", "norm"),
  norm_m = 0,
  norm_s = 1
)
}
\arguments{
\item{data}{a \link[dplyr:reexports]{dplyr::tibble} containing the (sub)scale item responses to be
scored}

\item{...}{<\link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select}> columns of the (sub)scale items in
order (See Note).}

\item{handinj}{<\link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select}> column specification of the
affected hand(s)}

\item{score}{desired score class}

\item{norm_m}{mean value to use when computing normative scores}

\item{norm_s}{standard deviation value to use when computing normative scores}
}
\value{
data augmented with score columns:
\enumerate{
\item mhq_asth_left_\if{html}{\out{<scale>}}
\itemize{
\item score from the left hand items
}
\item mhq_asth_right_\if{html}{\out{<scale>}}
\itemize{
\item score from the right hand items
}
\item mhq_asth_both_\if{html}{\out{<scale>}}
\itemize{
\item an average of the left and right scores
}
\item mhq_asth_\if{html}{\out{<scale>}}
\itemize{
\item the requested aesthetics scale score
}
}
}
\description{
MHQ Aesthetics
}
\note{
Items indicated in the ellipsis (...) are expected to be in the following
order:

V The following questions refer to the appearance (look) of your hands
during the past week.

A The following questions refer to the appearance (look) of your right hand.
\enumerate{
\item I was satisfied with the appearance (look) of my right hand.
\item The appearance (look) of my right hand sometimes made me uncomfortable in
public.
\item The appearance (look) of my right hand made me depressed.
\item The appearance (look) of my right hand interfered with my normal social
activities.
}

B The following questions refer to the appearance (look) of your left hand.
5. I was satisfied with the appearance (look) of my left hand.
6. The appearance (look) of my left hand sometimes made me uncomfortable in
public.
7. The appearance (look) of my left hand made me depressed.
8. The appearance (look) of my left hand interfered with my normal social
activities.
}
\examples{
mhq_asth(mhq, va1:vb4, handinj = handinj, score = "stnd")

}
