% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand.R
\name{mhq_hand}
\alias{mhq_hand}
\title{MHQ Overall Hand Function}
\usage{
mhq_hand(
  data,
  ...,
  handinj,
  score = c("raw", "mean", "stnd", "norm"),
  norm_m = 0,
  norm_s = 1
)
}
\arguments{
\item{data}{a \link[dplyr:reexports]{dplyr::tibble} containing the (sub)scale item responses to be
scored}

\item{...}{<\link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select}> columns of the (sub)scale items in
order (See Note).}

\item{handinj}{<\link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select}> column specification of the
affected hand(s)}

\item{score}{desired score class}

\item{norm_m}{mean value to use when computing normative scores}

\item{norm_s}{standard deviation value to use when computing normative scores}
}
\value{
data augmented with score columns:
\enumerate{
\item mhq_hand_left_\if{html}{\out{<score>}}
\itemize{
\item score from the left hand items
}
\item mhq_hand_right_\if{html}{\out{<score>}}
\itemize{
\item score from the right hand items
}
\item mhq_hand_both_\if{html}{\out{<score>}}
\itemize{
\item an average of the left and right hand scores
}
\item mhq_hand_\if{html}{\out{<score>}}
\itemize{
\item the requested overall hand function scale score
}
}
}
\description{
MHQ Overall Hand Function
}
\note{
Items indicated in the ellipsis (...) are expected to be in the following
order:

I The following questions refer to the function of your hand(s)/wrist(s)
during the past week.

A The following questions refer to your right hand/wrist.
\enumerate{
\item Overall, how well did your right hand work?
\item How well did your right fingers move?
\item How well did your right wrist move?
\item How was the strength in your right hand?
\item How was the sensation (feeling) in your right hand?
}

B The following questions refer to your left hand/wrist.
6. Overall, how well did your left hand work?
7. How well did your left fingers move?
8. How well did your left wrist move?
9. How was the strength in your left hand?
10. How was the sensation (feeling) in your left hand?
}
\examples{
mhq_hand(mhq, ia1:ib5, handinj = handinj, score = "stnd")

}
