% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{mhq_norms}
\alias{mhq_norms}
\title{MHQ Normative Values}
\usage{
mhq_norms(
  cohort = c("healthy", "trauma", "disease", "nonspecific", "systemic"),
  scale = c("hand", "aodl", "work", "pain", "asth", "sfac", "over"),
  score = c("raw", "mean", "stnd", "norm"),
  stat = c("m", "s")
)
}
\arguments{
\item{cohort}{cohort of patients from which values were calculated}

\item{scale}{desired (sub)scale}

\item{score}{desired score class}

\item{stat}{requested statistic}
}
\value{
the requested statistic or missing if not implemented
}
\description{
MHQ Normative Values
}
\examples{
mhq_norms("healthy", "hand", "stnd", "m")
}
