% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over.R
\name{mhq_over}
\alias{mhq_over}
\title{MHQ Overall}
\usage{
mhq_over(
  data,
  handinj,
  score = c("raw", "mean", "stnd", "norm"),
  norm_m = 0,
  norm_s = 1
)
}
\arguments{
\item{data}{a \link[dplyr:reexports]{dplyr::tibble} containing the (sub)scale item responses to be
scored}

\item{handinj}{<\link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select}> column specification of the
affected hand(s)}

\item{score}{desired score class}

\item{norm_m}{mean value to use when computing normative scores}

\item{norm_s}{standard deviation value to use when computing normative scores}
}
\value{
data augmented with score columns:
\enumerate{
\item mhq_overall_left_\if{html}{\out{<score>}}
\itemize{
\item overall score for the left hand
}
\item mhq_overall_right_\if{html}{\out{<score>}}
\itemize{
\item overall score for the right hand
}
\item mhq_overall_affected_\if{html}{\out{<score>}}
\itemize{
\item overall score for the affected hand(s)
}
\item mhq_overall_normal_\if{html}{\out{<score>}}
\itemize{
\item overall score for the unaffected hand
}
}
}
\description{
MHQ Overall
}
\examples{
mhq \%>\%
  mhq_hand(ia1:ib5, handinj = handinj, score = "stnd") \%>\%
  mhq_aodl(iia1:iic7, handinj = handinj, score = "stnd") \%>\%
  mhq_work(iii1:iii5, score = "stnd") \%>\%
  mhq_pain(iva1:ivb5, handinj = handinj, score = "stnd") \%>\%
  mhq_asth(va1:vb4, handinj = handinj, score = "stnd") \%>\%
  mhq_sfac(via1:vib6, handinj = handinj, score = "stnd") \%>\%
  mhq_over(handinj = handinj, score = "stnd")

}
\references{
Nolte, M. T., Shauver, M. J., & Chung, K. C. (2017). Normative values of the
Michigan Hand Outcomes Questionnaire for patients with and without hand
conditions. Plastic and reconstructive surgery, 140(3), 425e-433e, DOI:
10.1097/PRS.0000000000003581
}
