% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pain.R
\name{mhq_pain}
\alias{mhq_pain}
\title{MHQ Pain}
\usage{
mhq_pain(
  data,
  ...,
  handinj,
  score = c("raw", "mean", "stnd", "norm"),
  norm_m = 0,
  norm_s = 1
)
}
\arguments{
\item{data}{a \link[dplyr:reexports]{dplyr::tibble} containing the (sub)scale item responses to be
scored}

\item{...}{<\link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select}> columns of the (sub)scale items in
order (See Note).}

\item{handinj}{<\link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select}> column specification of the
affected hand(s)}

\item{score}{desired score class}

\item{norm_m}{mean value to use when computing normative scores}

\item{norm_s}{standard deviation value to use when computing normative scores}
}
\value{
data augmented with score columns:
\enumerate{
\item mhq_pain_left_\if{html}{\out{<scale>}}
\itemize{
\item score from the left hand items
}
\item mhq_pain_right_\if{html}{\out{<scale>}}
\itemize{
\item score from the right hand items
}
\item mhq_pain_both_\if{html}{\out{<scale>}}
\itemize{
\item an average of the left and right scores
}
\item mhq_pain_\if{html}{\out{<scale>}}
\itemize{
\item the requested pain scale score
}
}
}
\description{
MHQ Pain
}
\note{
Items indicated in the ellipsis (...) are expected to be in the following
order:

IV The following questions refer to how much pain you had in your
hand(s)/wrist(s) during the past week.

A The following questions refer to pain in your right hand/wrist.
\enumerate{
\item How often did you have pain in your right hand/wrist?
\item Please describe the pain you have in your right hand/wrist.
\item How often did the pain in your right hand/wrist interfere with your sleep?
\item How often did the pain in your right hand/wrist interfere with your daily
activities (such as eating or bathing)?
\item How often did the pain in your right hand/wrist make you unhappy?
}

B The following questions refer to pain in your left hand/wrist.
6. How often did you have pain in your left hand/wrist?
7. Please describe the pain you have in your left hand/wrist.
8. How often did the pain in your left hand/wrist interfere with your sleep?
9. How often did the pain in your left hand/wrist interfere with your daily
activities (such as eating or bathing)?
10. How often did the pain in your left hand/wrist make you unhappy?
}
\examples{
mhq_pain(mhq, iva1:ivb5, handinj = handinj, score = "stnd")

}
