% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo.R
\name{mhq_scoring_algorithm}
\alias{mhq_scoring_algorithm}
\title{Michigan Hand Outcomes Questionnaire (MHQ)}
\usage{
mhq_scoring_algorithm(
  data,
  scale = c("hand", "aodl", "work", "pain", "asth", "sfac"),
  reverse = TRUE,
  score = c("raw", "mean", "stnd", "norm"),
  norm_m,
  norm_s
)
}
\arguments{
\item{data}{a \link[dplyr:reexports]{dplyr::tibble} containing the (sub)scale item responses to be
scored}

\item{scale}{the intended (sub)scale}

\item{reverse}{logical indicating if the standardized and normalized scores
are to be reversed}

\item{score}{desired score class}

\item{norm_m}{mean value to use when computing normative scores}

\item{norm_s}{standard deviation value to use when computing normative scores}
}
\value{
a numeric vector of (sub)scale scores
}
\description{
Michigan Hand Outcomes Questionnaire (MHQ)
}
