% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work.R
\name{mhq_work}
\alias{mhq_work}
\title{MHQ Work}
\usage{
mhq_work(
  data,
  ...,
  score = c("raw", "mean", "stnd", "norm"),
  norm_m = 0,
  norm_s = 1
)
}
\arguments{
\item{data}{a \link[dplyr:reexports]{dplyr::tibble} containing the (sub)scale item responses to be
scored}

\item{...}{<\link[dplyr:dplyr_tidy_select]{dplyr::dplyr_tidy_select}> columns of the (sub)scale items in
order (See Note).}

\item{score}{desired score class}

\item{norm_m}{mean value to use when computing normative scores}

\item{norm_s}{standard deviation value to use when computing normative scores}
}
\value{
data augmented with mhq_work_\if{html}{\out{<scale>}}, the requested work scale score
}
\description{
MHQ Work
}
\note{
Items indicated in the ellipsis (...) are expected to be in the following
order:

III The following questions refer to how you did in your normal work
(including both housework and school work) during the past 4 weeks.
\enumerate{
\item How often were you unable to do your work because of problems with your
hand(s)/wrist(s)?
\item How often did you have to shorten your work day because of problems with
your hand(s)/wrist(s)?
\item How often did you have to take it easy at your work because of problems
with yoru hand(s)/wrist(s)?
\item How often did you accomplish less in your work because of problems with
your hand(s)/wrist(s)?
\item How often did you take longer to do the tasks in your work because of
problems with your hand(s)/wrist(s)?
}
}
\examples{
mhq_work(mhq, iii1:iii5, score = "stnd")

}
